# docs/v1/v1_swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ServiceAddGrpcRequest:
    properties:
      black_list:
        description: 黑名单
        example: ""
        type: string
      clientip_flow_limit:
        description: 客户端ip限流
        example: 0
        type: integer
      forbid_list:
        description: 禁用ip列表
        example: ""
        type: string
      header_transfor:
        description: 'header转换支持增加(add)、删除(del)、修改(edit) 格式: add headname headvalue
          多个逗号间隔'
        example: ""
        type: string
      ip_list:
        description: ip列表
        example: 47.113.203.197:2401
        type: string
      open_auth:
        description: service_access_control 表字段
        example: false
        type: boolean
      port:
        description: service_hrrp_rule 表字段
        type: integer
      round_type:
        description: service_load_balance 表字段
        example: 0
        type: integer
      service_desc:
        description: 服务描述
        example: 必填 服务描述
        type: string
      service_flow_limit:
        description: 服务的限流
        example: 0
        type: integer
      service_name:
        description: service_info 表字段
        example: 必填 服务名称
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        example: 0
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        example: 0
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        example: 0
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        example: 0
        type: integer
      weight_list:
        description: 权重列表
        example: "10"
        type: string
      white_list:
        description: 白名单
        example: ""
        type: string
    type: object
  dto.ServiceAddHttpRequest:
    properties:
      black_list:
        description: 黑名单
        example: ""
        type: string
      clientip_flow_limit:
        description: 客户端ip限流
        example: 0
        type: integer
      forbid_list:
        description: 禁用ip列表
        example: ""
        type: string
      header_transfor:
        description: 'header转换支持增加(add)、删除(del)、修改(edit) 格式: add headname headvalue
          多个逗号间隔'
        example: ""
        type: string
      ip_list:
        description: ip列表
        example: 47.113.203.197:2401
        type: string
      need_https:
        description: 支持https 1=支持
        example: false
        type: boolean
      need_strip_url:
        description: 启用strip_uri 1=启用
        example: false
        type: boolean
      need_websocket:
        description: 是否支持websocket 1=支持
        example: false
        type: boolean
      open_auth:
        description: service_access_control 表字段
        example: false
        type: boolean
      round_type:
        description: service_load_balance 表字段
        example: 0
        type: integer
      rule:
        description: type=domain表示域名，type=url_prefix时表示url前缀
        example: 必填 域名或后缀
        type: string
      rule_type:
        description: 匹配类型 0=url前缀url_prefix 1=域名domain
        example: 0
        type: integer
      service_desc:
        description: 服务描述
        example: 必填 服务描述
        type: string
      service_flow_limit:
        description: 服务的限流
        example: 0
        type: integer
      service_name:
        description: service_info 表字段
        example: 必填 服务名称
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        example: 0
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        example: 0
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        example: 0
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        example: 0
        type: integer
      url_rewrite:
        description: url重写功能 格式：^/gatekeeper/test_service(.*) $1 多个逗号间隔
        example: ""
        type: string
      weight_list:
        description: 权重列表
        example: "10"
        type: string
      white_list:
        description: 白名单
        example: ""
        type: string
    type: object
  dto.ServiceAddTcpRequest:
    properties:
      black_list:
        description: 黑名单
        example: ""
        type: string
      clientip_flow_limit:
        description: 客户端ip限流
        example: 0
        type: integer
      forbid_list:
        description: 禁用ip列表
        example: ""
        type: string
      ip_list:
        description: ip列表
        example: 47.113.203.197:2401
        type: string
      open_auth:
        description: service_access_control 表字段
        example: false
        type: boolean
      port:
        description: service_hrrp_rule 表字段
        type: integer
      round_type:
        description: service_load_balance 表字段
        example: 0
        type: integer
      service_desc:
        description: 服务描述
        example: 必填 服务描述
        type: string
      service_flow_limit:
        description: 服务的限流
        example: 0
        type: integer
      service_name:
        description: service_info 表字段
        example: 必填 服务名称
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        example: 0
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        example: 0
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        example: 0
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        example: 0
        type: integer
      weight_list:
        description: 权重列表
        example: "10"
        type: string
      white_list:
        description: 白名单
        example: ""
        type: string
    type: object
  dto.ServiceListItem:
    properties:
      id:
        type: integer
      load_type:
        type: integer
      qpd:
        type: integer
      qps:
        type: integer
      service_addr:
        type: string
      service_desc:
        type: string
      service_name:
        type: string
      total_node:
        type: integer
    type: object
  dto.ServiceListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.ServiceListItem'
        type: array
      total:
        type: integer
    type: object
...
  dto.ServiceStatResponse:
    properties:
      today:
        items:
          type: integer
        type: array
      yesterday:
        items:
          type: integer
        type: array
    type: object
  dto.ServiceUpdateGrpcRequest:
    properties:
      black_list:
        description: 黑名单
        example: ""
        type: string
      clientip_flow_limit:
        description: 客户端ip限流
        example: 0
        type: integer
      forbid_list:
        description: 禁用ip列表
        example: ""
        type: string
      header_transfor:
        description: 'header转换支持增加(add)、删除(del)、修改(edit) 格式: add headname headvalue
          多个逗号间隔'
        example: ""
        type: string
      id:
        type: integer
      ip_list:
        description: ip列表
        example: 47.113.203.197:2401
        type: string
      open_auth:
        description: service_access_control 表字段
        example: false
        type: boolean
      port:
        description: service_hrrp_rule 表字段
        type: integer
      round_type:
        description: service_load_balance 表字段
        example: 0
        type: integer
      service_desc:
        description: 服务描述
        example: 必填 服务描述
        type: string
      service_flow_limit:
        description: 服务的限流
        example: 0
        type: integer
      service_name:
        description: service_info 表字段
        example: 必填 服务名称
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        example: 0
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        example: 0
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        example: 0
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        example: 0
        type: integer
      weight_list:
        description: 权重列表
        example: "10"
        type: string
      white_list:
        description: 白名单
        example: ""
        type: string
    type: object
  dto.ServiceUpdateHttpRequest:
    properties:
      black_list:
        description: 黑名单
        example: ""
        type: string
      clientip_flow_limit:
        description: 客户端ip限流
        example: 0
        type: integer
      forbid_list:
        description: 禁用ip列表
        example: ""
        type: string
      header_transfor:
        description: 'header转换支持增加(add)、删除(del)、修改(edit) 格式: add headname headvalue
          多个逗号间隔'
        example: ""
        type: string
      id:
        type: integer
      ip_list:
        description: ip列表
        example: 47.113.203.197:2401
        type: string
      need_https:
        description: 支持https 1=支持
        example: false
        type: boolean
      need_strip_url:
        description: 启用strip_uri 1=启用
        example: false
        type: boolean
      need_websocket:
        description: 是否支持websocket 1=支持
        example: false
        type: boolean
      open_auth:
        description: service_access_control 表字段
        example: false
        type: boolean
      round_type:
        description: service_load_balance 表字段
        example: 0
        type: integer
      rule:
        description: type=domain表示域名，type=url_prefix时表示url前缀
        example: 必填 域名或后缀
        type: string
      rule_type:
        description: 匹配类型 0=url前缀url_prefix 1=域名domain
        example: 0
        type: integer
      service_desc:
        description: 服务描述
        example: 必填 服务描述
        type: string
      service_flow_limit:
        description: 服务的限流
        example: 0
        type: integer
      service_name:
        description: service_info 表字段
        example: 必填 服务名称
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        example: 0
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        example: 0
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        example: 0
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        example: 0
        type: integer
      url_rewrite:
        description: url重写功能 格式：^/gatekeeper/test_service(.*) $1 多个逗号间隔
        example: ""
        type: string
      weight_list:
        description: 权重列表
        example: "10"
        type: string
      white_list:
        description: 白名单
        example: ""
        type: string
    type: object
  dto.ServiceUpdateTcpRequest:
    properties:
      black_list:
        description: 黑名单
        example: ""
        type: string
      clientip_flow_limit:
        description: 客户端ip限流
        example: 0
        type: integer
      forbid_list:
        description: 禁用ip列表
        example: ""
        type: string
      id:
        type: integer
      ip_list:
        description: ip列表
        example: 47.113.203.197:2401
        type: string
      open_auth:
        description: service_access_control 表字段
        example: false
        type: boolean
      port:
        description: service_hrrp_rule 表字段
        type: integer
      round_type:
        description: service_load_balance 表字段
        example: 0
        type: integer
      service_desc:
        description: 服务描述
        example: 必填 服务描述
        type: string
      service_flow_limit:
        description: 服务的限流
        example: 0
        type: integer
      service_name:
        description: service_info 表字段
        example: 必填 服务名称
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        example: 0
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        example: 0
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        example: 0
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        example: 0
        type: integer
      weight_list:
        description: 权重列表
        example: "10"
        type: string
      white_list:
        description: 白名单
        example: ""
        type: string
    type: object
  middleware.ResponseErr:
    properties:
      data: {}
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: string
    type: object
host: 127.0.0.1:2401
info:
  contact:
    email: stone2401@qq.com
  description: 一个轻量级网关代理，支持多种代理协议，接口转发，数据统计，带有管理界面
  title: 轻量级网关
  version: "1.0"
paths:
...
  /service/service_add_grpc:
    post:
      consumes:
      - application/json
      description: 添加hGrpc服务
      operationId: /service/service_add_grpc
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddGrpcRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 添加hGrpc服务
      tags:
      - 服务管理
  /service/service_add_http:
    post:
      consumes:
      - application/json
      description: 添加http服务
      operationId: /service/service_add_http
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddHttpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 添加http服务
      tags:
      - 服务管理
  /service/service_add_tcp:
    post:
      consumes:
      - application/json
      description: 添加tcp服务
      operationId: /service/service_add_tcp
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddTcpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 添加tcp服务
      tags:
      - 服务管理
  /service/service_delete:
    get:
      consumes:
      - application/json
      description: 服务删除
      operationId: /service/service_delete
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 关键词
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 服务删除
      tags:
      - 服务管理
  /service/service_detail:
    get:
      consumes:
      - application/json
      description: 服务详情
      operationId: /service/service_detail
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 关键词
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  $ref: '#/definitions/dto.ServiceUpdateGrpcRequest'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 服务详情
      tags:
      - 服务管理
  /service/service_list:
    get:
      consumes:
      - application/json
      description: 服务列表
      operationId: /service/service_list
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 页数
        in: query
        name: page_no
        required: true
        type: integer
      - description: 个数
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  $ref: '#/definitions/dto.ServiceListResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 服务列表
      tags:
      - 服务管理
  /service/service_stat:
    get:
      consumes:
      - application/json
      description: 服务统计
      operationId: /service/service_stat
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 关键词
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  $ref: '#/definitions/dto.ServiceStatResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 服务统计
      tags:
      - 服务管理
  /service/service_update_grpc:
    post:
      consumes:
      - application/json
      description: 修改tcp服务
      operationId: /service/service_update_grpc
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateGrpcRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改grpc服务
      tags:
      - 服务管理
  /service/service_update_http:
    post:
      consumes:
      - application/json
      description: 修改http服务
      operationId: /service/service_update_http
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateHttpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改http服务
      tags:
      - 服务管理
  /service/service_update_tcp:
    post:
      consumes:
      - application/json
      description: 修改tcp服务
      operationId: /service/service_update_tcp
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateTcpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.ResponseErr'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改tcp服务
      tags:
      - 服务管理
schemes:
- http
swagger: "2.0"